Validate valid ECR URL in refreshECRToken() by calling isECRregistry()
